Update for x/tools/go/packages
Update string2enum so that it works in the presence of go modules and without GOPATH.

Process I took to make this patch:

1) Identified the fork commit from x/tools
2) Condensed down to a single patch, modifying as little as possible (undid some unimportant syntactic changes, renamed back to cmd/stringer)
3) Applied that patch to the fork commit in x/tools
4) Did a `git imerge rebase` onto x/tools origin/master
5) Fixed some minor broken things
6) Used the import path information now present from go/packages (pkgPath)
7) Took the new stringer.go and copied it back into this repository

The new fork commit in x/tools is:

```diff
commit 5a8dccf5b48ae388eebac31f4077f14adda4dbc6 (origin/master, origin/HEAD, master)
Author: Ian Cottrell <iancottrell@google.com>
Date:   Fri Mar 22 16:43:20 2019 -0400

    go/packages/packagestest: convert to the span library

    Change-Id: Ib8984dd0828e3bb0b49eaa1604649300a4d54306
    Reviewed-on: https://go-review.googlesource.com/c/tools/+/168900
    Reviewed-by: Michael Matloob <matloob@golang.org>
```

This process can be repeated in the future. Just copy this file as it is in this tree over the file in the other tree, making a single patch. Then imerge rebase it, which should be a straightforward process.

The one disadvantage with how I've done this is that the history of the other repository isn't preserved in this repository. But the effort for that is too large for the benefit, I think. The patch is only small, anyway!

What I found is that the new `string2enum` "just worked", and no longer needs the goutil package. It additionally introduces the new sanity check which was added in stringer.